fix(exec): Validate in-memory exchange buffer sizing
Preserve maxExchangeBufferSize as uint64_t in the transport context.
Reject values above int64_t before constructing the in-memory client.
Name exchange.max_buffer_size in the user error and cover the boundary.

+++ velox/exec/Task.cpp
@@ -21,6 +21,8 @@
 #include <exception>
 #include <string>
 
+#include <folly/Conv.h>
+
 #include "velox/common/base/Counters.h"
 #include "velox/common/base/StatsReporter.h"
 #include "velox/common/file/FileSystems.h"
@@ -3901,8 +3903,7 @@ void Task::createExchangeClientLocked(
           .taskId = taskId_,
           .destination = destination_,
           .numberOfConsumers = numberOfConsumers,
-          .maxExchangeBufferSize =
-              static_cast<int64_t>(queryConfig.maxExchangeBufferSize()),
+          .maxExchangeBufferSize = queryConfig.maxExchangeBufferSize(),
           .minExchangeOutputBatchBytes =
               queryConfig.minExchangeOutputBatchBytes(),
           .pool = addExchangeClientPool(planNodeId, pipelineId),
